Add an example flashrd role
This is all I needed to do now that the flashrd.site patches are in
flashrd's github repo.

This was as simple as `echo role-flashrd >> common/siteXX/roles`
and it just worked.  (I am sure that some things also needed to be
adjusted for the readonly filesystems but it boots up)