A somewhat working example of nsh for amd64
The compiled binaries are compiled for amd64, not that I wanted to,
but I haven't come up with a good way to deploy specific binaries
for different architectures/versions.

The rc.firsttime script works fairly well, needs to disable the
roles that it is taking over for, but haven't come up with a good
solution for that.